rename trackhandlingutils to trackcontentuploadmanager
