+++ creator-node/src/components/tracks/tracksComponentService.js
@@ -4,7 +4,7 @@ const {
   getStartTime
 } = require('../../logging')
 
-const TrackHandlingUtils = require('./trackHandlingUtils')
+const TrackContentUploadManager = require('./trackContentUploadManager')
 
 /**
  * Upload track segment files and make avail - will later be associated with Audius track
@@ -33,7 +33,7 @@ const handleTrackContentRoute = async ({ logContext }, requestProps) => {
   // Create track transcode and segments, and save all to disk
   const codeBlockTimeStart = getStartTime()
   const { transcodeFilePath, segmentFileNames } =
-    await TrackHandlingUtils.transcodeAndSegment(
+    await TrackContentUploadManager.transcodeAndSegment(
       { logContext },
       { fileName, fileDir }
     )
@@ -42,7 +42,7 @@ const handleTrackContentRoute = async ({ logContext }, requestProps) => {
     `Successfully re-encoded track file=${fileName}`
   )
 
-  const resp = await TrackHandlingUtils.processTranscodeAndSegments(
+  const resp = await TrackContentUploadManager.processTranscodeAndSegments(
     { logContext },
     {
       session: { cnodeUserUUID },
@@ -65,7 +65,7 @@ async function handleTranscodeAndSegment(
   { logContext },
   { fileName, fileDir }
 ) {
-  return TrackHandlingUtils.transcodeAndSegment(
+  return TrackContentUploadManager.transcodeAndSegment(
     { logContext },
     { fileName, fileDir }
   )
